requirements.txt
aiofiles==25.1.0
asyncpg==0.31.0
celery[redis]==5.6.3
fastapi==0.135.3
geoalchemy2==0.18.4
httpx==0.28.1
minio==7.2.20
numpy==2.4.4
openai==2.31.0
passlib[bcrypt]==1.7.4
Pillow==12.2.0
piexif==1.1.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
python-multipart==0.0.24
python-socketio==5.16.1
redis==6.1.1
sqlalchemy[asyncio]==2.0.49
structlog==25.5.0
uvicorn[standard]==0.44.0
twilio==9.4.0